Refactor get_placeholder_with_slot to be in utils.placeholder
